+++ packages/nexusgraph-nlp/index.ts
@@ -1,3 +1,3 @@
 // Copyright 2023 Paion Data. All rights reserved.
-export type { NLPClient as NaturalLanguageProcessor } from "./src/processor/NLPClient";
-export { TheresaClient as RemoteNaturalLanguageProcessor } from "./src/processor/TheresaClient";
+export type { NLPClient as NaturalLanguageProcessor } from "./src/NLPClient";
+export { TheresaClient as RemoteNaturalLanguageProcessor } from "./src/TheresaClient";
