+++ packages/nexusgraph-nlp/src/TheresaClient.test.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 Paion Data. All rights reserved.
 import axios from "axios";
 
-import { container, TYPES } from "../../../nexusgraph-app/inversify.config";
+import { container, TYPES } from "../../nexusgraph-app/inversify.config";
 import { NLPClient } from "./NLPClient";
 
 const remoteNaturalLanguageProcessor: any = container.get<NLPClient>(TYPES.NaturalLanguageProcessor);
