--- packages/nexusgraph-nlp/src/parser/index.ts
@@ -1,2 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-export { default as LexicalEditorStateParser } from "./TextareaContentParser";
